+++ src/Controller/Enemy/DeleteController.php
@@ -5,6 +5,7 @@ namespace App\Controller\Enemy;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use App\Repository\EnemyRepository;
 
 final class DeleteController extends AbstractController
@@ -18,6 +19,11 @@ final class DeleteController extends AbstractController
         int $id,
     ): JsonResponse {
         $enemy = $this->enemyRepo->find($id);
+
+        if (!$enemy) {
+            throw new NotFoundHttpException('Enemy not found');
+        }
+
         $this->enemyRepo->delete($enemy);
 
         return new JsonResponse(null, 204);
