+++ src/Controller/Knight/DeleteController.php
@@ -5,6 +5,7 @@ namespace App\Controller\Knight;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use App\Repository\KnightRepository;
 
 final class DeleteController extends AbstractController
@@ -18,6 +19,11 @@ final class DeleteController extends AbstractController
         int $id,
     ): JsonResponse {
         $knight = $this->knightRepo->find($id);
+
+        if (!$knight) {
+            throw new NotFoundHttpException('Knight not found');
+        }
+
         $this->knightRepo->delete($knight);
 
         return new JsonResponse(null, 204);
