+++ src/Controller/LootPool/DeleteController.php
@@ -5,6 +5,7 @@ namespace App\Controller\LootPool;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\HttpKernel\Exception\NotFoundHttpException;
 use App\Repository\LootPoolRepository;
 
 final class DeleteController extends AbstractController
@@ -18,6 +19,11 @@ final class DeleteController extends AbstractController
         int $id,
     ): JsonResponse {
         $lootPool = $this->lootPoolRepo->find($id);
+
+        if (!$lootPool) {
+            throw new NotFoundHttpException('Loot pool not found');
+        }
+
         $this->lootPoolRepo->delete($lootPool);
 
         return new JsonResponse(null, 204);
